% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTMLDocument.R
\name{as.character.HTMLDocument}
\alias{as.character.HTMLDocument}
\title{Retrieve character representation of HTMLElement}
\usage{
\method{as.character}{HTMLDocument}(x, include_declaration = TRUE, ...)
}
\arguments{
\item{x}{HTMLElement object}

\item{include_declaration}{Include the HTML5 DOCTYPE declaration at the top? Default: TRUE}

\item{...}{other arguments}
}
\description{
Retrieve character representation of HTMLElement
}
