% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTMLElement.R
\name{as.character.HTMLElement}
\alias{as.character.HTMLElement}
\title{Retrieve character representation of HTMLElement}
\usage{
\method{as.character}{HTMLElement}(x, include_declaration = FALSE, ...)
}
\arguments{
\item{x}{HTMLElement object}

\item{include_declaration}{Include the HTML5 DOCTYPE declaration at the top? Default: FALSE}

\item{...}{other arguments}
}
\description{
Retrieve character representation of HTMLElement
}
