% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htag.R
\name{htag}
\alias{htag}
\title{HTML Builder functions}
\description{
HTML builder functions.  Similar in purpose to \code{shiny::tags},
but with auto-complete of attribute names as part of the function call. Also
based upon \code{HTMLElement} at its core (rather than a list, as in \code{shiny})

\code{htag} is a dreadful portmanteau of \emph{html} and \emph{tag}.

The simpler name of \code{tag} was rejected to avoid a name clash with the
\code{shiny} package.

Wrapper functions (around \code{HTMLElement$new()}) to assist in building
HTML documents
}
\examples{
\dontrun{
htag$a(href = 'hello', "greeting")
# <a href="hello">greeting</a>
}


}
