% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTMLDocument.R
\docType{data}
\name{HTMLDocument}
\alias{HTMLDocument}
\alias{html_doc}
\title{HTMLDocument Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HTMLDocument
}
\description{
This is a slightly specialized subclass of \code{HTMLElement} which includes
a \code{$show()} method, and includes the HTML DOCTYPE by default.
}
\keyword{datasets}
