% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTMLElement.R
\name{knit_print.HTMLElement}
\alias{knit_print.HTMLElement}
\title{knitr/rmarkdown compatibility/output}
\usage{
knit_print.HTMLElement(x, ...)
}
\arguments{
\item{x}{HTMLElement}

\item{...}{other arguments}
}
\description{
knitr/rmarkdown compatibility/output
}
