% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTMLDocument.R
\name{HTMLDocument}
\alias{HTMLDocument}
\alias{html_doc}
\title{HTMLDocument Class}
\description{
HTMLDocument Class

HTMLDocument Class
}
\details{
This is a slightly specialized subclass of \code{HTMLElement} which includes
a \code{$show()} method, and includes the HTML DOCTYPE by default.

only tested in MacOS and Rstudio
}
\section{Super class}{
\code{\link[minihtml:HTMLElement]{minihtml::HTMLElement}} -> \code{HTMLDocument}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HTMLDocument$new()}}
\item \href{#method-save}{\code{HTMLDocument$save()}}
\item \href{#method-print}{\code{HTMLDocument$print()}}
\item \href{#method-as_character}{\code{HTMLDocument$as_character()}}
\item \href{#method-show}{\code{HTMLDocument$show()}}
}
}
\if{html}{\subsection{Inherited methods}{
\itemize{
\item \href{../../minihtml/html/HTMLElement.html#method-a}{\code{minihtml::HTMLElement$a()}}
\item \href{../../minihtml/html/HTMLElement.html#method-abbr}{\code{minihtml::HTMLElement$abbr()}}
\item \href{../../minihtml/html/HTMLElement.html#method-add}{\code{minihtml::HTMLElement$add()}}
\item \href{../../minihtml/html/HTMLElement.html#method-address}{\code{minihtml::HTMLElement$address()}}
\item \href{../../minihtml/html/HTMLElement.html#method-append}{\code{minihtml::HTMLElement$append()}}
\item \href{../../minihtml/html/HTMLElement.html#method-area}{\code{minihtml::HTMLElement$area()}}
\item \href{../../minihtml/html/HTMLElement.html#method-article}{\code{minihtml::HTMLElement$article()}}
\item \href{../../minihtml/html/HTMLElement.html#method-as_character_inner}{\code{minihtml::HTMLElement$as_character_inner()}}
\item \href{../../minihtml/html/HTMLElement.html#method-aside}{\code{minihtml::HTMLElement$aside()}}
\item \href{../../minihtml/html/HTMLElement.html#method-audio}{\code{minihtml::HTMLElement$audio()}}
\item \href{../../minihtml/html/HTMLElement.html#method-b}{\code{minihtml::HTMLElement$b()}}
\item \href{../../minihtml/html/HTMLElement.html#method-base}{\code{minihtml::HTMLElement$base()}}
\item \href{../../minihtml/html/HTMLElement.html#method-bdi}{\code{minihtml::HTMLElement$bdi()}}
\item \href{../../minihtml/html/HTMLElement.html#method-bdo}{\code{minihtml::HTMLElement$bdo()}}
\item \href{../../minihtml/html/HTMLElement.html#method-big}{\code{minihtml::HTMLElement$big()}}
\item \href{../../minihtml/html/HTMLElement.html#method-blockquote}{\code{minihtml::HTMLElement$blockquote()}}
\item \href{../../minihtml/html/HTMLElement.html#method-body}{\code{minihtml::HTMLElement$body()}}
\item \href{../../minihtml/html/HTMLElement.html#method-br}{\code{minihtml::HTMLElement$br()}}
\item \href{../../minihtml/html/HTMLElement.html#method-button}{\code{minihtml::HTMLElement$button()}}
\item \href{../../minihtml/html/HTMLElement.html#method-canvas}{\code{minihtml::HTMLElement$canvas()}}
\item \href{../../minihtml/html/HTMLElement.html#method-caption}{\code{minihtml::HTMLElement$caption()}}
\item \href{../../minihtml/html/HTMLElement.html#method-cite}{\code{minihtml::HTMLElement$cite()}}
\item \href{../../minihtml/html/HTMLElement.html#method-code}{\code{minihtml::HTMLElement$code()}}
\item \href{../../minihtml/html/HTMLElement.html#method-col}{\code{minihtml::HTMLElement$col()}}
\item \href{../../minihtml/html/HTMLElement.html#method-colgroup}{\code{minihtml::HTMLElement$colgroup()}}
\item \href{../../minihtml/html/HTMLElement.html#method-command}{\code{minihtml::HTMLElement$command()}}
\item \href{../../minihtml/html/HTMLElement.html#method-copy}{\code{minihtml::HTMLElement$copy()}}
\item \href{../../minihtml/html/HTMLElement.html#method-datalist}{\code{minihtml::HTMLElement$datalist()}}
\item \href{../../minihtml/html/HTMLElement.html#method-dd}{\code{minihtml::HTMLElement$dd()}}
\item \href{../../minihtml/html/HTMLElement.html#method-del}{\code{minihtml::HTMLElement$del()}}
\item \href{../../minihtml/html/HTMLElement.html#method-details}{\code{minihtml::HTMLElement$details()}}
\item \href{../../minihtml/html/HTMLElement.html#method-dfn}{\code{minihtml::HTMLElement$dfn()}}
\item \href{../../minihtml/html/HTMLElement.html#method-dialog}{\code{minihtml::HTMLElement$dialog()}}
\item \href{../../minihtml/html/HTMLElement.html#method-div}{\code{minihtml::HTMLElement$div()}}
\item \href{../../minihtml/html/HTMLElement.html#method-dl}{\code{minihtml::HTMLElement$dl()}}
\item \href{../../minihtml/html/HTMLElement.html#method-dt}{\code{minihtml::HTMLElement$dt()}}
\item \href{../../minihtml/html/HTMLElement.html#method-em}{\code{minihtml::HTMLElement$em()}}
\item \href{../../minihtml/html/HTMLElement.html#method-embed}{\code{minihtml::HTMLElement$embed()}}
\item \href{../../minihtml/html/HTMLElement.html#method-fieldset}{\code{minihtml::HTMLElement$fieldset()}}
\item \href{../../minihtml/html/HTMLElement.html#method-figcaption}{\code{minihtml::HTMLElement$figcaption()}}
\item \href{../../minihtml/html/HTMLElement.html#method-figure}{\code{minihtml::HTMLElement$figure()}}
\item \href{../../minihtml/html/HTMLElement.html#method-footer}{\code{minihtml::HTMLElement$footer()}}
\item \href{../../minihtml/html/HTMLElement.html#method-form}{\code{minihtml::HTMLElement$form()}}
\item \href{../../minihtml/html/HTMLElement.html#method-h1}{\code{minihtml::HTMLElement$h1()}}
\item \href{../../minihtml/html/HTMLElement.html#method-h2}{\code{minihtml::HTMLElement$h2()}}
\item \href{../../minihtml/html/HTMLElement.html#method-h3}{\code{minihtml::HTMLElement$h3()}}
\item \href{../../minihtml/html/HTMLElement.html#method-h4}{\code{minihtml::HTMLElement$h4()}}
\item \href{../../minihtml/html/HTMLElement.html#method-h5}{\code{minihtml::HTMLElement$h5()}}
\item \href{../../minihtml/html/HTMLElement.html#method-h6}{\code{minihtml::HTMLElement$h6()}}
\item \href{../../minihtml/html/HTMLElement.html#method-head}{\code{minihtml::HTMLElement$head()}}
\item \href{../../minihtml/html/HTMLElement.html#method-header}{\code{minihtml::HTMLElement$header()}}
\item \href{../../minihtml/html/HTMLElement.html#method-hgroup}{\code{minihtml::HTMLElement$hgroup()}}
\item \href{../../minihtml/html/HTMLElement.html#method-hr}{\code{minihtml::HTMLElement$hr()}}
\item \href{../../minihtml/html/HTMLElement.html#method-html}{\code{minihtml::HTMLElement$html()}}
\item \href{../../minihtml/html/HTMLElement.html#method-i}{\code{minihtml::HTMLElement$i()}}
\item \href{../../minihtml/html/HTMLElement.html#method-iframe}{\code{minihtml::HTMLElement$iframe()}}
\item \href{../../minihtml/html/HTMLElement.html#method-ilayer}{\code{minihtml::HTMLElement$ilayer()}}
\item \href{../../minihtml/html/HTMLElement.html#method-img}{\code{minihtml::HTMLElement$img()}}
\item \href{../../minihtml/html/HTMLElement.html#method-input}{\code{minihtml::HTMLElement$input()}}
\item \href{../../minihtml/html/HTMLElement.html#method-ins}{\code{minihtml::HTMLElement$ins()}}
\item \href{../../minihtml/html/HTMLElement.html#method-kbd}{\code{minihtml::HTMLElement$kbd()}}
\item \href{../../minihtml/html/HTMLElement.html#method-keygen}{\code{minihtml::HTMLElement$keygen()}}
\item \href{../../minihtml/html/HTMLElement.html#method-label}{\code{minihtml::HTMLElement$label()}}
\item \href{../../minihtml/html/HTMLElement.html#method-legend}{\code{minihtml::HTMLElement$legend()}}
\item \href{../../minihtml/html/HTMLElement.html#method-li}{\code{minihtml::HTMLElement$li()}}
\item \href{../../minihtml/html/HTMLElement.html#method-link}{\code{minihtml::HTMLElement$link()}}
\item \href{../../minihtml/html/HTMLElement.html#method-main}{\code{minihtml::HTMLElement$main()}}
\item \href{../../minihtml/html/HTMLElement.html#method-map}{\code{minihtml::HTMLElement$map()}}
\item \href{../../minihtml/html/HTMLElement.html#method-mark}{\code{minihtml::HTMLElement$mark()}}
\item \href{../../minihtml/html/HTMLElement.html#method-marquee}{\code{minihtml::HTMLElement$marquee()}}
\item \href{../../minihtml/html/HTMLElement.html#method-menu}{\code{minihtml::HTMLElement$menu()}}
\item \href{../../minihtml/html/HTMLElement.html#method-meta}{\code{minihtml::HTMLElement$meta()}}
\item \href{../../minihtml/html/HTMLElement.html#method-meter}{\code{minihtml::HTMLElement$meter()}}
\item \href{../../minihtml/html/HTMLElement.html#method-nav}{\code{minihtml::HTMLElement$nav()}}
\item \href{../../minihtml/html/HTMLElement.html#method-noscript}{\code{minihtml::HTMLElement$noscript()}}
\item \href{../../minihtml/html/HTMLElement.html#method-object}{\code{minihtml::HTMLElement$object()}}
\item \href{../../minihtml/html/HTMLElement.html#method-ol}{\code{minihtml::HTMLElement$ol()}}
\item \href{../../minihtml/html/HTMLElement.html#method-optgroup}{\code{minihtml::HTMLElement$optgroup()}}
\item \href{../../minihtml/html/HTMLElement.html#method-option}{\code{minihtml::HTMLElement$option()}}
\item \href{../../minihtml/html/HTMLElement.html#method-output}{\code{minihtml::HTMLElement$output()}}
\item \href{../../minihtml/html/HTMLElement.html#method-p}{\code{minihtml::HTMLElement$p()}}
\item \href{../../minihtml/html/HTMLElement.html#method-param}{\code{minihtml::HTMLElement$param()}}
\item \href{../../minihtml/html/HTMLElement.html#method-pre}{\code{minihtml::HTMLElement$pre()}}
\item \href{../../minihtml/html/HTMLElement.html#method-progress}{\code{minihtml::HTMLElement$progress()}}
\item \href{../../minihtml/html/HTMLElement.html#method-q}{\code{minihtml::HTMLElement$q()}}
\item \href{../../minihtml/html/HTMLElement.html#method-remove}{\code{minihtml::HTMLElement$remove()}}
\item \href{../../minihtml/html/HTMLElement.html#method-rp}{\code{minihtml::HTMLElement$rp()}}
\item \href{../../minihtml/html/HTMLElement.html#method-rt}{\code{minihtml::HTMLElement$rt()}}
\item \href{../../minihtml/html/HTMLElement.html#method-ruby}{\code{minihtml::HTMLElement$ruby()}}
\item \href{../../minihtml/html/HTMLElement.html#method-samp}{\code{minihtml::HTMLElement$samp()}}
\item \href{../../minihtml/html/HTMLElement.html#method-script}{\code{minihtml::HTMLElement$script()}}
\item \href{../../minihtml/html/HTMLElement.html#method-section}{\code{minihtml::HTMLElement$section()}}
\item \href{../../minihtml/html/HTMLElement.html#method-select}{\code{minihtml::HTMLElement$select()}}
\item \href{../../minihtml/html/HTMLElement.html#method-small}{\code{minihtml::HTMLElement$small()}}
\item \href{../../minihtml/html/HTMLElement.html#method-source}{\code{minihtml::HTMLElement$source()}}
\item \href{../../minihtml/html/HTMLElement.html#method-span}{\code{minihtml::HTMLElement$span()}}
\item \href{../../minihtml/html/HTMLElement.html#method-strong}{\code{minihtml::HTMLElement$strong()}}
\item \href{../../minihtml/html/HTMLElement.html#method-style}{\code{minihtml::HTMLElement$style()}}
\item \href{../../minihtml/html/HTMLElement.html#method-sub}{\code{minihtml::HTMLElement$sub()}}
\item \href{../../minihtml/html/HTMLElement.html#method-summary}{\code{minihtml::HTMLElement$summary()}}
\item \href{../../minihtml/html/HTMLElement.html#method-sup}{\code{minihtml::HTMLElement$sup()}}
\item \href{../../minihtml/html/HTMLElement.html#method-table}{\code{minihtml::HTMLElement$table()}}
\item \href{../../minihtml/html/HTMLElement.html#method-tbody}{\code{minihtml::HTMLElement$tbody()}}
\item \href{../../minihtml/html/HTMLElement.html#method-td}{\code{minihtml::HTMLElement$td()}}
\item \href{../../minihtml/html/HTMLElement.html#method-template}{\code{minihtml::HTMLElement$template()}}
\item \href{../../minihtml/html/HTMLElement.html#method-textarea}{\code{minihtml::HTMLElement$textarea()}}
\item \href{../../minihtml/html/HTMLElement.html#method-tfoot}{\code{minihtml::HTMLElement$tfoot()}}
\item \href{../../minihtml/html/HTMLElement.html#method-th}{\code{minihtml::HTMLElement$th()}}
\item \href{../../minihtml/html/HTMLElement.html#method-thead}{\code{minihtml::HTMLElement$thead()}}
\item \href{../../minihtml/html/HTMLElement.html#method-time}{\code{minihtml::HTMLElement$time()}}
\item \href{../../minihtml/html/HTMLElement.html#method-title}{\code{minihtml::HTMLElement$title()}}
\item \href{../../minihtml/html/HTMLElement.html#method-tr}{\code{minihtml::HTMLElement$tr()}}
\item \href{../../minihtml/html/HTMLElement.html#method-track}{\code{minihtml::HTMLElement$track()}}
\item \href{../../minihtml/html/HTMLElement.html#method-tt}{\code{minihtml::HTMLElement$tt()}}
\item \href{../../minihtml/html/HTMLElement.html#method-ul}{\code{minihtml::HTMLElement$ul()}}
\item \href{../../minihtml/html/HTMLElement.html#method-update}{\code{minihtml::HTMLElement$update()}}
\item \href{../../minihtml/html/HTMLElement.html#method-var}{\code{minihtml::HTMLElement$var()}}
\item \href{../../minihtml/html/HTMLElement.html#method-video}{\code{minihtml::HTMLElement$video()}}
\item \href{../../minihtml/html/HTMLElement.html#method-wbr}{\code{minihtml::HTMLElement$wbr()}}
}
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialise HTMLDocument
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HTMLDocument$new(..., name = "html")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Further arguments passed to \code{HTMLElement$new()}}

\item{\code{name}}{name of node. default: 'html'}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\subsection{Method \code{save()}}{
Save HTML fragment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HTMLDocument$save(filename, include_declaration = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{filename}

\item{\code{include_declaration}}{Include \code{DOCTYPE} declaration? default: TRUE}

\item{\code{...}}{Further arguments passed to \code{HTMLElement$save()}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print HTMLDocument
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HTMLDocument$print(include_declaration = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_declaration}}{Include \code{DOCTYPE} declaration? default: TRUE}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_character"></a>}}
\subsection{Method \code{as_character()}}{
Convert to character representation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HTMLDocument$as_character(include_declaration = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_declaration}}{Include \code{DOCTYPE} declaration? default: TRUE}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
single character string representing the HTML
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\subsection{Method \code{show()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HTMLDocument$show(viewer = getOption("viewer", utils::browseURL))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{viewer}}{viewer to use}
}
\if{html}{\out{</div>}}
}
}
}
